Fix bug with Eventual
Previously if the function that an Eventual bound caused the Eventual to
reschedule, it would deadlock. When I removed the unneeded locking code,
it then caused the Eventual to not reschedule properly. As it turns out,
calling the reschedule on the Eventual triggered killing the current
waiting task, which (because we did not spawn the bound function) would
kill the greenlet calling the reschedule.

We now properly spawn a new greenlet to call the bound function, which
avoids any previously existing binding.