[admin] add the ability to temporarily place users in arbitrary roles
This commit adds support for temporarily placing users in arbitrary
roles, similar to the way tempmute works. This was requested by
r/Overwatch, and is simple enough to append onto infractions. There is
still some cleanup to be had here, and there are bits that are fragile
to invalid roles in aliases.