[utilities] add async queries, use in !info
Previously queries within functions like !info where executed
synchronously and sequentially. This is actually _very_ inefficient
because most of the time we spend on any query is in the network or DB,
meaning time gevent actually just unschedules for. To resolve this I've
added support to peewee to kick off the query execution within a
greenlet, and then return an AsyncResult. My theory is this will heavily
improve the performance of the !info command, but I'm also tracking
query latencies in the function to see if we have a slow/bad query.