[admin] merge mute/tempmute to fix use case oddities
One of the most common mistakes I've seen (and done myself) when using
mute/tempmute is to call normal mute with a duration. This commit merges
mute and tempmute and adds some smart support to mute for checking
whether the first part of the reason message can actually be parsed as a
duration. This should fix most of the woes people have with this
command (and also reduces code duplication a tiny bit)